backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.6
python-dotenv==1.0.1
azure-cosmos==4.6.0
pydantic==2.7.1
